# Conflicts:
#	Domain/UserInput/MapField.cs
#	UI-MVC/Areas/Identity/Pages/Account/Manage/_LayoutRegular.cshtml
#	UI-MVC/Startup.cs
#	UI-MVC/TheSpanishDatabase.db
#	UI-MVC/Views/Platform/CollectProject.cshtml
#	UI-MVC/appsettings.json

+++ Domain/UserInput/MapField.cs
@@ -2,7 +2,7 @@ namespace Domain.UserInput
 {
     public class MapField : Field
     {
-        public float LocationX { get; set; }
-        public float LocationY { get; set; }    
+        public double LocationX { get; set; }
+        public double LocationY { get; set; }
     }
 }
